requirements.txt
faster-whisper==1.0.3
ffmpeg-python==0.2.0
python-dotenv==1.2.2
requests==2.32.5